Harden WooCommerce policy and settings isolation
